New Update
the last feature of this update is the copy of the dashboard from the internet site of Formula 1.
So far the design & layout structure are already in what could be a finished version.
The tables of the first 2 (out of 7) tabs related to specific races are done. The tables of the remaing 5 tabs still need to be created